add interface for BookController
