+++ hw09-spring_mvc/src/test/java/ru/otus/controllers/BookControllerImplTest.java
@@ -20,8 +20,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  * @author Ilya Rogatkin
  */
 @DisplayName("Контроллер BookController должен")
-@WebMvcTest(BookController.class)
-class BookControllerTest {
+@WebMvcTest(BookControllerImpl.class)
+class BookControllerImplTest {
 
     @Autowired
     private MockMvc mockMvc;
